+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### ⚠ Breaking
 - require Jenkins 2.375.2 or newer
 - require Java 11 or newer ([required since Jenkins 2.361.1](https://www.jenkins.io/blog/2022/06/28/require-java-11/))
+- require Dependency-Track 4.7 or newer
 
 ### ⭐ New Features
 ### 🐞 Bugs Fixed